reformated Blocklisten.md to make it work with jacklul/pihole-updatelists
jacklul/pihole-updatelists (https://github.com/jacklul/pihole-updatelists) is a tool that allows to automatically import newly added lists into the Pi-hole. For it to work, it needs to be formatted correctly. So I removed the asterisk before the URLs and put the "comments" in separate lines and commented it out.